feat(playground): implement playground editor and output panels
