+++ src/components/playground/constants.ts
@@ -1,0 +1,3 @@
+export const STORAGE_KEY = 'js-playground-code'
+export const DEFAULT_SNIPPET = `console.log('Hello from the playground!')
+`
